Rename romaji.js to matcher.js describing its purpose better
